Split test methods out of the Molecule class.
